commit con proyecto en kotlin

+++ app/src/main/java/com/example/proyecto/models/InvoiceVO.kt
@@ -1,5 +1,7 @@
 package com.example.proyecto.models
 
-data class InvoiceVO( private final  var descEstado:String,
-                      private final  var importeOrdenacion:Long ,
-                      private  final var fecha:String)
+data class InvoiceVO(
+    private final var descEstado: String,
+    private final var importeOrdenacion: Double,
+    private final var fecha: String
+)
